+++ cookiecutter.json
@@ -2,25 +2,26 @@
   "full_name": "First Middle Last",
   "email": "email@example.com",
   "autora_contribution_type": ["theorist", "experimentalist", "experiment_runner", "synthetic_data", "custom"],
-  "custom_autora_contribution_type": "{% if cookiecutter.autora_contribution_type == 'custom' -%}<custom_autora_contribution_type>{% else -%}NOT APPLICABLE{% endif -%}",
+  "custom_autora_contribution_type": "{% if cookiecutter.autora_contribution_type == 'custom' -%}<custom_autora_contribution_type>{% else -%}N/A - Press Enter to Skip{% endif -%}",
   "__autora_contribution_type": "{% if cookiecutter.autora_contribution_type == 'custom' -%}{{ cookiecutter.custom_autora_contribution_type }}{% else -%}{{ cookiecutter.autora_contribution_type }}{% endif -%}",
   "experimentalist_type": [
-    "{% if cookiecutter.autora_contribution_type == 'experimentalist' -%}sampler{% else -%}NOT APPLICABLE{% endif -%}", 
-    "{% if cookiecutter.autora_contribution_type == 'experimentalist' -%}pooler{% else -%}NOT APPLICABLE{% endif -%}", 
-    "{% if cookiecutter.autora_contribution_type == 'experimentalist' -%}custom{% else -%}NOT APPLICABLE{% endif -%}"],
-  "custom_experimentalist_type": "{% if cookiecutter.experimentalist_type == 'custom' -%}<custom_experimentalist_type>{% else -%}NOT APPLICABLE{% endif -%}",
+    "{% if cookiecutter.autora_contribution_type == 'experimentalist' -%}sampler{% else -%}N/A - Press Enter to Skip{% endif -%}", 
+    "{% if cookiecutter.autora_contribution_type == 'experimentalist' -%}pooler{% else -%}N/A - Press Enter to Skip{% endif -%}", 
+    "{% if cookiecutter.autora_contribution_type == 'experimentalist' -%}custom{% else -%}N/A - Press Enter to Skip{% endif -%}"],
+  "custom_experimentalist_type": "{% if cookiecutter.experimentalist_type == 'custom' -%}<custom_experimentalist_type>{% else -%}N/A - Press Enter to Skip{% endif -%}",
   "__experimentalist_type": "{% if cookiecutter.autora_contribution_type == 'experimentalist' and cookiecutter.experimentalist_type == 'custom' -%}{{ cookiecutter.custom_experimentalist_type }}{% else -%}{{ cookiecutter.experimentalist_type }}{% endif -%}",
   "experiement_runner_type": [
-    "{% if cookiecutter.autora_contribution_type == 'experiment_runner' -%}experiment_runner{% else -%}NOT APPLICABLE{% endif -%}", 
-    "{% if cookiecutter.autora_contribution_type == 'experiment_runner' -%}participation_manager{% else -%}NOT APPLICABLE{% endif -%}", 
-    "{% if cookiecutter.autora_contribution_type == 'experiment_runner' -%}recruitment_manager{% else -%}NOT APPLICABLE{% endif -%}"],
-  "__contribution_subtype": "{% if cookiecutter.__autora_contribution_type == 'experimentalist' -%}{{ cookiecutter.__experimentalist_type }}{% elif cookiecutter.__autora_contribution_type == 'experiment_runner' and cookiecutter.experiement_runner_type != 'experiment_runner' -%}{{ cookiecutter.experiement_runner_type }}{% endif -%}",
+    "{% if cookiecutter.autora_contribution_type == 'experiment_runner' -%}experiment_runner{% else -%}N/A - Press Enter to Skip{% endif -%}", 
+    "{% if cookiecutter.autora_contribution_type == 'experiment_runner' -%}participation_manager{% else -%}N/A - Press Enter to Skip{% endif -%}", 
+    "{% if cookiecutter.autora_contribution_type == 'experiment_runner' -%}recruitment_manager{% else -%}N/A - Press Enter to Skip{% endif -%}"],
+  "__contribution_subtype": "{% if cookiecutter.__autora_contribution_type == 'experimentalist' -%}{{ cookiecutter.__experimentalist_type }}{% elif cookiecutter.__autora_contribution_type == 'experiment_runner' and cookiecutter.experiement_runner_type != 'experiment_runner' -%}{{ cookiecutter.experiement_runner_type }}{% else -%}not-applicable{% endif -%}",
   "contribution_name": "<contribution name>",
   "project_short_description": "<Short description>",
   "repository": "<Repository URL>",
   "use_dynamic_versioning": ["yes", "no"],
   "use_github_actions": ["yes", "no"],
   "license": ["MIT license", "BSD license", "ISC license", "Apache Software License 2.0", "GNU General Public License v3", "Not open source"],
-  "__project_slug": "autora-{{ cookiecutter.autora_contribution_type }}-{% if cookiecutter.__contribution_subtype != '' -%}{{ cookiecutter.__contribution_subtype }}-{% endif -%}{{ cookiecutter.contribution_name.lower().replace(' ', '-') }}",
+  "__full_path": "{{ cookiecutter.autora_contribution_type }}-{% if cookiecutter.__contribution_subtype != 'not-applicable' -%}{{ cookiecutter.__contribution_subtype }}-{% endif -%}{{ cookiecutter.contribution_name.lower().replace(' ', '-') }}",
+  "__project_slug": "autora-{{ cookiecutter.__full_path }}",
   "__python_name": "{{ cookiecutter.contribution_name.lower().replace(' ', '_').replace('-', '_') }}"
 }
