+++ {{ cookiecutter.__project_slug }}/docs/quickstart.md
@@ -4,10 +4,10 @@ You will need:
 
 - `python` 3.8 or greater: [https://www.python.org/downloads/](https://www.python.org/downloads/)
 
-*Your-contribution* is a part of the `autora` package:
+*{{ cookiecutter.contribution_name }} is a part of the `autora` package:
 
 ```shell
-pip install -U autora["your-contribution"]
+pip install -U autora["{{ cookiecutter.__autora_contribution_type }}"]
 ```
