--- {{ cookiecutter.__project_slug }}/src/autora/experiment_runner/experimentation_manager/{{ cookiecutter.__python_name }}/__init__.py
@@ -1,31 +1,0 @@
-"""
-Example experimentation manager:
-    These are example functions that could be included 
-    (but are not necessarily included)
-"""
-
-
-def send_conditions(args, conditions):
-    """
-    Add a description here
-    Args:
-        args: the arguments needed to set up the experiment (e.g., credentials)
-        conditions: the conditions to be sent
-    """
-    pass
-
-
-def check_status(args):
-    """
-    Add description here
-    """
-    pass
-
-
-def get_observations(args):
-    """
-    Add description here
-    """
-    observations = None
-    return observations
-
