--- {{ cookiecutter.__project_slug }}/src/autora/experiment_runner/recruitment_manager/{{ cookiecutter.__python_name }}/__init__.py
@@ -1,48 +1,0 @@
-"""
-Example recruitment manager:
-    These are example functions that could be included 
-    (but are not necessarily included)
-"""
-
-
-def example_setup_recruitment(args):
-    """
-    Add your description here.
-    Set up a study on an example recruitment platform.
-    
-    Args:
-        the arguments needed to set up the study (e.g., credentials)
-    """
-    pass
-
-
-def start_recruitment(args):
-    """
-    Add description here.
-    Start the recruitment on an example recruitment platform.
-
-    Args:
-        the arguments needed to start the recruitment (e.g., credentials)
-    """
-    pass
-
-
-def pause_recruitment(args):
-    """
-    Add description here.
-    Pause the recruitment on an example recruitment platform.
-
-    Args:
-        the arguments needed to start the recruitment (e.g., credentials)
-    """
-
-
-def check_recruitment_status(args):
-    """
-    Add description here.
-    Get information on the status of recruitment for a study on an example recruitment platform.
-    
-    Args:
-        the arguments needed to start the recruitment (e.g., credentials)
-    """
-    pass
