--- {{ cookiecutter.__project_slug }}/src/autora/experiment_runner/{{ cookiecutter.__python_name }}/__init__.py
@@ -1,17 +1,0 @@
-"""
-Example Experiment Runner
-"""
-
-
-def example_runner(conditions):
-    """
-    Add a description of the experiment runner here
-
-    Args:
-        conditions: list of conditions to get observations for
-    Returns: observations
-
-    """
-    observations = conditions
-
-    return observations
